Crate cosmic_protocols

This crate provides bindings to the COSMIC wayland protocol extensions.

These bindings are built on top of the crates wayland-client and wayland-server.

Each protocol module contains a client and a server submodules, for each side of the protocol. The creation of these modules (and the dependency on the associated crate) is controlled by the two cargo features client and server.

Modules

a11y

Accessibility support.

corner_radius

Hint toplevel corner radius values.

image_capture_source

Capture source interface extending ext-image-capture-source-v1.

output_management

Output management interface.

overlap_notify

Get overlap notifications for layer surfaces

toplevel_info

Receive information about toplevel surfaces.

toplevel_management

Modify state toplevel surfaces.

workspace

Receive information about and control workspaces.