cosmic_protocols::output_management::v1::server
Module zcosmic_output_head_v1

Output extension object

Extension to zwlr_output_head_v1.

Adds additional read-only properties.

Properties sent via this interface are applied atomically via the wlr_output_manager.done event. No guarantees are made regarding the order in which properties are sent.

Structs

ZcosmicOutputHeadV1

Output extension object

Enums

AdaptiveSyncAvailability

AdaptiveSyncStateExt

Event

Request

Constants

EVT_ADAPTIVE_SYNC_AVAILABLE_OPCODE

The wire opcode for this event

EVT_ADAPTIVE_SYNC_AVAILABLE_SINCE

The minimal object version supporting this event

EVT_ADAPTIVE_SYNC_EXT_OPCODE

The wire opcode for this event

EVT_ADAPTIVE_SYNC_EXT_SINCE

The minimal object version supporting this event

EVT_MIRRORING_OPCODE

The wire opcode for this event

EVT_MIRRORING_SINCE

The minimal object version supporting this event

EVT_SCALE_1000_OPCODE

The wire opcode for this event

EVT_SCALE_1000_SINCE

The minimal object version supporting this event

REQ_RELEASE_OPCODE

The wire opcode for this request

REQ_RELEASE_SINCE

The minimal object version supporting this request