Enum cosmic_protocols::screencopy::v1::client::zcosmic_screencopy_session_v1::Event

#[non_exhaustive]
pub enum Event {
    BufferInfo {
        _type: WEnum<BufferType>,
        node: Option<String>,
        format: u32,
        width: u32,
        height: u32,
        stride: u32,
    },
    InitDone,
    Transform {
        transform: WEnum<Transform>,
    },
    Damage {
        x: u32,
        y: u32,
        width: u32,
        height: u32,
    },
    CursorEnter {
        seat: WlSeat,
        input_type: WEnum<InputType>,
    },
    CursorLeave {
        seat: WlSeat,
        input_type: WEnum<InputType>,
    },
    CursorInfo {
        seat: WlSeat,
        input_type: WEnum<InputType>,
        position_x: i32,
        position_y: i32,
        width: i32,
        height: i32,
        hotspot_x: i32,
        hotspot_y: i32,
    },
    Failed {
        reason: WEnum<FailureReason>,
    },
    CommitTime {
        tv_sec_hi: u32,
        tv_sec_lo: u32,
        tv_nsec: u32,
    },
    Ready,
}

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

BufferInfo

Fields

_type: WEnum<BufferType>

buffer type

node: Option<String>

device node

format: u32

buffer drm format

width: u32

buffer width

height: u32

buffer height

stride: u32

buffer stride

buffer information

Provides information about buffer parameters that need to be used for the main image. This event is sent for every supported buffer type after the session is created.

The stride parameter is invalid for dmabuf and may be set to 0.

InitDone

session initialisation done

This event is sent once when all buffer info events have been sent.

Transform

Fields

transform: WEnum<Transform>

carries the transform

This event is sent before the ready event and holds transformations of the source buffer.

Damage

Fields

x: u32

damaged x coordinates

y: u32

damaged y coordinates

width: u32

current width

height: u32

current height

carries the coordinates of the damaged region

This event is sent before the ready event. It may be generated multiple times for each commit.

The arguments describe a box around an area that has changed since the last ready event.

These coordinates originate in the upper left corner of the buffer.

CursorEnter

Fields

seat: WlSeat

input_type: WEnum<InputType>

input type

cursor entered surface

Sent when a cursor enters the captured surface. It shall be generated before the “cursor_info” event when and only when a cursor enters the surface.

CursorLeave

Fields

seat: WlSeat

input_type: WEnum<InputType>

input type

cursor left surface

Sent when a cursor leaves the captured surface. No “cursor_info” event is generated for for the given cursor.

CursorInfo

Fields

seat: WlSeat

input_type: WEnum<InputType>

position_x: i32

position x coordinates

position_y: i32

position y coordinates

width: i32

width of the cursor image

height: i32

height of the cursor image

hotspot_x: i32

hotspot x coordinates

hotspot_y: i32

hotspot y coordinates

cursor specific information

This event is generated for each cursor buffer that was attached to the session and for which the cursor is currently focused in the session. It is generated once for each cursor buffer before the ready event.

Cursors outside the surface do not get captured and no event will be generated for them.

If the cursor image has changed, the cursor buffer will have been updated and the “has_damage” argument will be set to 1; otherwise 0.

The given position is the position of the cursor’s hotspot and it is relative to the main buffer’s top left corner in transformed buffer pixel coordinates.

The hotspot coordinates are relative to the cursor buffers upper left corner.

Failed

Fields

reason: WEnum<FailureReason>

commit failed

This event indicates that the attempted frame copy has failed.

After receiving this event, the client must destroy the object.

CommitTime

Fields

tv_sec_hi: u32

high 32 bits of the seconds part of the timestamp

tv_sec_lo: u32

low 32 bits of the seconds part of the timestamp

tv_nsec: u32

nanoseconds part of the timestamp

indicates the commit time of the frame

This event indicates the time at which the frame is committed to be scanned out in system monotonic time.

The timestamp is expressed as tv_sec_hi, tv_sec_lo, tv_nsec triples, each component being an unsigned 32-bit value. Whole seconds are in tv_sec which is a 64-bit value combined from tv_sec_hi and tv_sec_lo, and the additional fractional part in tv_nsec as nanoseconds. Hence, for valid timestamps tv_nsec must be in [0, 999999999].

Ready

indicates frame is available for reading

Called as soon as the frame is copied, indicating it is available for reading.

Implementations

impl Event

pub fn opcode(&self) -> u16

Get the opcode number of this message