Module cosmic_protocols::screencopy::v2::client::zcosmic_screencopy_session_v2

screen capture session

This object represents an active screencopy session.

After a screencopy session is created, buffer constraint events will be emitted from the compositor to tell the client which buffer types and formats are supported for reading from the session. The compositor may re-send buffer constraint events whenever they change.

The advertise buffer constraints, the compositor must send in no particular order: zero or more shm_format and dmabuf_format events, zero or one dmabuf_device event, and exactly one buffer_size event. Then the compositor must send a done event.

When the client has received all the buffer constraints, it can create a buffer accordingly, attach it to the screencopy session using the attach_buffer request, set the buffer damage using the damage_buffer request and then send the capture request.

Structs

ZcosmicScreencopySessionV2

screen capture session

Enums

Event

Request

Constants

EVT_BUFFER_SIZE_OPCODE

The wire opcode for this event

EVT_BUFFER_SIZE_SINCE

The minimal object version supporting this event

EVT_DMABUF_DEVICE_OPCODE

The wire opcode for this event

EVT_DMABUF_DEVICE_SINCE

The minimal object version supporting this event

EVT_DMABUF_FORMAT_OPCODE

The wire opcode for this event

EVT_DMABUF_FORMAT_SINCE

The minimal object version supporting this event

EVT_DONE_OPCODE

The wire opcode for this event

EVT_DONE_SINCE

The minimal object version supporting this event

EVT_SHM_FORMAT_OPCODE

The wire opcode for this event

EVT_SHM_FORMAT_SINCE

The minimal object version supporting this event

EVT_STOPPED_OPCODE

The wire opcode for this event

EVT_STOPPED_SINCE

The minimal object version supporting this event

REQ_CREATE_FRAME_OPCODE

The wire opcode for this request

REQ_CREATE_FRAME_SINCE

The minimal object version supporting this request

REQ_DESTROY_OPCODE

The wire opcode for this request

REQ_DESTROY_SINCE

The minimal object version supporting this request