Module cosmic_protocols::screencopy::v2::server::zcosmic_screencopy_frame_v2

screen capture frame

This object represents a screen capture frame.

The client should attach a buffer, damage the buffer, and then send a capture request.

If the screen capture is successful, the compositor will send the frame metadata (transform, damage, presentation_time in any order) followed by the ready event.

If the screen capture fails, the compositor will send the failed event.

Structs

ZcosmicScreencopyFrameV2

screen capture frame

Enums

Error

Event

FailureReason

Request

Constants

EVT_DAMAGE_OPCODE

The wire opcode for this event

EVT_DAMAGE_SINCE

The minimal object version supporting this event

EVT_FAILED_OPCODE

The wire opcode for this event

EVT_FAILED_SINCE

The minimal object version supporting this event

EVT_PRESENTATION_TIME_OPCODE

The wire opcode for this event

EVT_PRESENTATION_TIME_SINCE

The minimal object version supporting this event

EVT_READY_OPCODE

The wire opcode for this event

EVT_READY_SINCE

The minimal object version supporting this event

EVT_TRANSFORM_OPCODE

The wire opcode for this event

EVT_TRANSFORM_SINCE

The minimal object version supporting this event

REQ_ATTACH_BUFFER_OPCODE

The wire opcode for this request

REQ_ATTACH_BUFFER_SINCE

The minimal object version supporting this request

REQ_CAPTURE_OPCODE

The wire opcode for this request

REQ_CAPTURE_SINCE

The minimal object version supporting this request

REQ_DAMAGE_BUFFER_OPCODE

The wire opcode for this request

REQ_DAMAGE_BUFFER_SINCE

The minimal object version supporting this request

REQ_DESTROY_OPCODE

The wire opcode for this request

REQ_DESTROY_SINCE

The minimal object version supporting this request