Module cosmic_protocols::toplevel_info::v1::server::zcosmic_toplevel_handle_v1

an open toplevel

A zcosmic_toplevel_handle_v1 object represents an open toplevel window. A single app may have multiple open toplevels.

Each toplevel has a list of outputs it is visible on, exposed to the client via the output_enter and output_leave events.

Structs

ZcosmicToplevelHandleV1

an open toplevel

Enums

Event

Request

State

types of states on the toplevel

Constants

EVT_APP_ID_OPCODE

The wire opcode for this event

EVT_APP_ID_SINCE

The minimal object version supporting this event

EVT_CLOSED_OPCODE

The wire opcode for this event

EVT_CLOSED_SINCE

The minimal object version supporting this event

EVT_DONE_OPCODE

The wire opcode for this event

EVT_DONE_SINCE

The minimal object version supporting this event

EVT_OUTPUT_ENTER_OPCODE

The wire opcode for this event

EVT_OUTPUT_ENTER_SINCE

The minimal object version supporting this event

EVT_OUTPUT_LEAVE_OPCODE

The wire opcode for this event

EVT_OUTPUT_LEAVE_SINCE

The minimal object version supporting this event

EVT_STATE_OPCODE

The wire opcode for this event

EVT_STATE_SINCE

The minimal object version supporting this event

EVT_TITLE_OPCODE

The wire opcode for this event

EVT_TITLE_SINCE

The minimal object version supporting this event

EVT_WORKSPACE_ENTER_OPCODE

The wire opcode for this event

EVT_WORKSPACE_ENTER_SINCE

The minimal object version supporting this event

EVT_WORKSPACE_LEAVE_OPCODE

The wire opcode for this event

EVT_WORKSPACE_LEAVE_SINCE

The minimal object version supporting this event

REQ_DESTROY_OPCODE

The wire opcode for this request

REQ_DESTROY_SINCE

The minimal object version supporting this request