Module cosmic_protocols::toplevel_management::v1::server::zcosmic_toplevel_manager_v1

control open apps

This protocol allows clients such as a taskbar to request the compositor to preform typical actions on open toplevels. The compositor is in all cases free to ignore the request.

Structs

ZcosmicToplevelManagerV1

control open apps

Enums

Error

Event

Request

ZcosmicToplelevelManagementCapabilitiesV1

Constants

EVT_CAPABILITIES_OPCODE

The wire opcode for this event

EVT_CAPABILITIES_SINCE

The minimal object version supporting this event

REQ_ACTIVATE_OPCODE

The wire opcode for this request

REQ_ACTIVATE_SINCE

The minimal object version supporting this request

REQ_CLOSE_OPCODE

The wire opcode for this request

REQ_CLOSE_SINCE

The minimal object version supporting this request

REQ_DESTROY_OPCODE

The wire opcode for this request

REQ_DESTROY_SINCE

The minimal object version supporting this request

REQ_SET_FULLSCREEN_OPCODE

The wire opcode for this request

REQ_SET_FULLSCREEN_SINCE

The minimal object version supporting this request

REQ_SET_MAXIMIZED_OPCODE

The wire opcode for this request

REQ_SET_MAXIMIZED_SINCE

The minimal object version supporting this request

REQ_SET_MINIMIZED_OPCODE

The wire opcode for this request

REQ_SET_MINIMIZED_SINCE

The minimal object version supporting this request

REQ_SET_RECTANGLE_OPCODE

The wire opcode for this request

REQ_SET_RECTANGLE_SINCE

The minimal object version supporting this request

REQ_UNSET_FULLSCREEN_OPCODE

The wire opcode for this request

REQ_UNSET_FULLSCREEN_SINCE

The minimal object version supporting this request

REQ_UNSET_MAXIMIZED_OPCODE

The wire opcode for this request

REQ_UNSET_MAXIMIZED_SINCE

The minimal object version supporting this request

REQ_UNSET_MINIMIZED_OPCODE

The wire opcode for this request

REQ_UNSET_MINIMIZED_SINCE

The minimal object version supporting this request