cosmic_protocols::toplevel_management::v1::server::zcosmic_toplevel_manager_v1

Enum Request

#[non_exhaustive]
pub enum Request {

    Destroy,
    Close {
        toplevel: ZcosmicToplevelHandleV1,
    },
    Activate {
        toplevel: ZcosmicToplevelHandleV1,
        seat: WlSeat,
    },
    SetMaximized {
        toplevel: ZcosmicToplevelHandleV1,
    },
    UnsetMaximized {
        toplevel: ZcosmicToplevelHandleV1,
    },
    SetMinimized {
        toplevel: ZcosmicToplevelHandleV1,
    },
    UnsetMinimized {
        toplevel: ZcosmicToplevelHandleV1,
    },
    SetFullscreen {
        toplevel: ZcosmicToplevelHandleV1,
        output: Option<WlOutput>,
    },
    UnsetFullscreen {
        toplevel: ZcosmicToplevelHandleV1,
    },
    SetRectangle {
        toplevel: ZcosmicToplevelHandleV1,
        surface: WlSurface,
        x: i32,
        y: i32,
        width: i32,
        height: i32,
    },
    MoveToWorkspace {
        toplevel: ZcosmicToplevelHandleV1,
        workspace: ZcosmicWorkspaceHandleV1,
        output: WlOutput,
    },
    SetSticky {
        toplevel: ZcosmicToplevelHandleV1,
    },
    UnsetSticky {
        toplevel: ZcosmicToplevelHandleV1,
    },
    MoveToExtWorkspace {
        toplevel: ZcosmicToplevelHandleV1,
        workspace: ExtWorkspaceHandleV1,
        output: WlOutput,
    },

}

Available on crate feature server only.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Destroy

destroy the zcosmic_toplevel_manager_v1

This request indicates that the client has finished using the zcosmic_toplevel_manager_v1 object and that it can be safely destroyed.

This is a destructor, once received this object cannot be used any longer.

Close

request that a toplevel be closed

If the compositor honors this request, the zcosmic_toplevel_handle_v1.closed event will be sent.

Fields

toplevel: ZcosmicToplevelHandleV1

Activate

request that a toplevel be activated

If the compositor honors this request, the zcosmic_toplevel_handle_v1.state event will be sent.

Fields

toplevel: ZcosmicToplevelHandleV1

seat: WlSeat

SetMaximized

request that a toplevel be maximized

If the compositor honors this request, the zcosmic_toplevel_handle_v1.state event will be sent.

Fields

toplevel: ZcosmicToplevelHandleV1

UnsetMaximized

request that a toplevel be unmaximized

If the compositor honors this request, the zcosmic_toplevel_handle_v1.state event will be sent.

Fields

toplevel: ZcosmicToplevelHandleV1

SetMinimized

request that a toplevel be minimized

If the compositor honors this request, the zcosmic_toplevel_handle_v1.state event will be sent.

Fields

toplevel: ZcosmicToplevelHandleV1

UnsetMinimized

request that a toplevel be unminimized

If the compositor honors this request, the zcosmic_toplevel_handle_v1.state event will be sent.

Fields

toplevel: ZcosmicToplevelHandleV1

SetFullscreen

request that a toplevel be fullscreened

If the compositor honors this request, the zcosmic_toplevel_handle_v1.state and potentially the zcosmic_toplevel_handle_v1.output_enter/output_leave events will be sent.

The output parameter a hint to the compositor and may be ignored. A value of NULL indicates that the compositor should choose the target output, if it honors the fullscreen request.

Fields

toplevel: ZcosmicToplevelHandleV1

output: Option<WlOutput>

UnsetFullscreen

request that a toplevel be unfullscreened

If the compositor honors this request, the zcosmic_toplevel_handle_v1.state event will be sent.

Fields

toplevel: ZcosmicToplevelHandleV1

SetRectangle

designate a rectangle to represent a toplevel

If a client using this protocol displays UI elements corresponding to toplevels, it may use this request to inform the server about such a relation. This information may be used by the server, for example as the target for a minimize animation.

If the client sets more than one rectangle, only the most recently set rectangle is considered.

The dimensions are given in surface-local coordinates.

Setting width=height=0 removes the current rectangle if one was set.

Fields

toplevel: ZcosmicToplevelHandleV1

surface: WlSurface

x: i32

y: i32

width: i32

height: i32

MoveToWorkspace

move toplevel to workspace

Move window to workspace, on given output.

Only available since version 2 of the interface

Fields

toplevel: ZcosmicToplevelHandleV1

workspace: ZcosmicWorkspaceHandleV1

output: WlOutput

SetSticky

request that a toplevel be made sticky

If the compositor honors this request, the zcosmic_toplevel_handle_v1.state event will be sent.

Only available since version 3 of the interface

Fields

toplevel: ZcosmicToplevelHandleV1

UnsetSticky

request that a toplevel be removed of the sticky status

If the compositor honors this request, the zcosmic_toplevel_handle_v1.state event will be sent.

Only available since version 3 of the interface

Fields

toplevel: ZcosmicToplevelHandleV1

MoveToExtWorkspace

move toplevel to workspace

Move window to workspace, on given output.

Only available since version 4 of the interface

Fields

toplevel: ZcosmicToplevelHandleV1

workspace: ExtWorkspaceHandleV1

output: WlOutput

Implementations

impl Request

pub fn opcode(&self) -> u16

Get the opcode number of this message