Module cosmic_protocols::screencopy::v1::client::zcosmic_screencopy_session_v1

a capture session

This object represents a session that’s being captured.

After a screencopy session is created, buffer_info events will be emitted from the compositor to tell the client which buffer types and formats are supported for reading from the surface.

When the client knows all the buffer attributes, it can create a buffer, attach it to the screencopy surface using the “attach_buffer” request, set the buffer damage using the “damage_buffer” request and then call the “commit” request.

After “commit” has been called, the next time that a buffer is committed by the compositor, the contents of that buffer will be copied to the one committed to the screencopy session. A series of events will be generated, ending with the “ready” event, which means that the buffer is ready to be used and a buffer may be committed to the surface again.

The “failed” event may be sent at any time. When this happens, the client must destroy the session. Depending on the failure reason, the client can create a new session to replace it.

Structs

Options

ZcosmicScreencopySessionV1

a capture session

Enums

BufferType

Error

Event

FailureReason

InputType

Request

Constants

EVT_BUFFER_INFO_OPCODE

The wire opcode for this event

EVT_BUFFER_INFO_SINCE

The minimal object version supporting this event

EVT_COMMIT_TIME_OPCODE

The wire opcode for this event

EVT_COMMIT_TIME_SINCE

The minimal object version supporting this event

EVT_CURSOR_ENTER_OPCODE

The wire opcode for this event

EVT_CURSOR_ENTER_SINCE

The minimal object version supporting this event

EVT_CURSOR_INFO_OPCODE

The wire opcode for this event

EVT_CURSOR_INFO_SINCE

The minimal object version supporting this event

EVT_CURSOR_LEAVE_OPCODE

The wire opcode for this event

EVT_CURSOR_LEAVE_SINCE

The minimal object version supporting this event

EVT_DAMAGE_OPCODE

The wire opcode for this event

EVT_DAMAGE_SINCE

The minimal object version supporting this event

EVT_FAILED_OPCODE

The wire opcode for this event

EVT_FAILED_SINCE

The minimal object version supporting this event

EVT_INIT_DONE_OPCODE

The wire opcode for this event

EVT_INIT_DONE_SINCE

The minimal object version supporting this event

EVT_READY_OPCODE

The wire opcode for this event

EVT_READY_SINCE

The minimal object version supporting this event

EVT_TRANSFORM_OPCODE

The wire opcode for this event

EVT_TRANSFORM_SINCE

The minimal object version supporting this event

REQ_ATTACH_BUFFER_OPCODE

The wire opcode for this request

REQ_ATTACH_BUFFER_SINCE

The minimal object version supporting this request

REQ_CAPTURE_CURSOR_OPCODE

The wire opcode for this request

REQ_CAPTURE_CURSOR_SINCE

The minimal object version supporting this request

REQ_COMMIT_OPCODE

The wire opcode for this request

REQ_COMMIT_SINCE

The minimal object version supporting this request

REQ_DESTROY_OPCODE

The wire opcode for this request

REQ_DESTROY_SINCE

The minimal object version supporting this request