Module cosmic_protocols::workspace::v1::client::zcosmic_workspace_handle_v1

a workspace handing a group of surfaces

A zcosmic_workspace_handle_v1 object represents a a workspace that handles a group of surfaces.

Each workspace has a name, conveyed to the client with the name event; a list of states, conveyed to the client with the state event; and optionally a set of coordinates, conveyed to the client with the coordinates event. The client may request that the compositor activate or deactivate the workspace.

Each workspace can belong to only a single workspace group. Depepending on the compositor policy, there might be workspaces with the same name in different workspace groups, but these workspaces are still separate (e.g. one of them might be active while the other is not).

Structs

ZcosmicWorkspaceHandleV1

a workspace handing a group of surfaces

Enums

Event

Request

State

types of states on the workspace

ZcosmicWorkspaceCapabilitiesV1

Constants

EVT_CAPABILITIES_OPCODE

The wire opcode for this event

EVT_CAPABILITIES_SINCE

The minimal object version supporting this event

EVT_COORDINATES_OPCODE

The wire opcode for this event

EVT_COORDINATES_SINCE

The minimal object version supporting this event

EVT_NAME_OPCODE

The wire opcode for this event

EVT_NAME_SINCE

The minimal object version supporting this event

EVT_REMOVE_OPCODE

The wire opcode for this event

EVT_REMOVE_SINCE

The minimal object version supporting this event

EVT_STATE_OPCODE

The wire opcode for this event

EVT_STATE_SINCE

The minimal object version supporting this event

REQ_ACTIVATE_OPCODE

The wire opcode for this request

REQ_ACTIVATE_SINCE

The minimal object version supporting this request

REQ_DEACTIVATE_OPCODE

The wire opcode for this request

REQ_DEACTIVATE_SINCE

The minimal object version supporting this request

REQ_DESTROY_OPCODE

The wire opcode for this request

REQ_DESTROY_SINCE

The minimal object version supporting this request

REQ_REMOVE_OPCODE

The wire opcode for this request

REQ_REMOVE_SINCE

The minimal object version supporting this request