Module cosmic_protocols::workspace::v1::server::zcosmic_workspace_group_handle_v1

a workspace group assigned to a set of outputs

A zcosmic_workspace_group_handle_v1 object represents a a workspace group that is assigned a set of outputs and contains a number of workspaces.

The set of outputs assigned to the workspace group is conveyed to the client via output_enter and output_leave events, and its workspaces are conveyed with workspace events.

For example, a compositor which has a set of workspaces for each output may advertise a workspace group (and its workspaces) per output, whereas a compositor where a workspace spans all outputs may advertise a single workspace group for all outputs.

Structs

ZcosmicWorkspaceGroupHandleV1

a workspace group assigned to a set of outputs

Enums

Event

Request

ZcosmicWorkspaceGroupCapabilitiesV1

Constants

EVT_CAPABILITIES_OPCODE

The wire opcode for this event

EVT_CAPABILITIES_SINCE

The minimal object version supporting this event

EVT_OUTPUT_ENTER_OPCODE

The wire opcode for this event

EVT_OUTPUT_ENTER_SINCE

The minimal object version supporting this event

EVT_OUTPUT_LEAVE_OPCODE

The wire opcode for this event

EVT_OUTPUT_LEAVE_SINCE

The minimal object version supporting this event

EVT_REMOVE_OPCODE

The wire opcode for this event

EVT_REMOVE_SINCE

The minimal object version supporting this event

EVT_WORKSPACE_OPCODE

The wire opcode for this event

EVT_WORKSPACE_SINCE

The minimal object version supporting this event

REQ_CREATE_WORKSPACE_OPCODE

The wire opcode for this request

REQ_CREATE_WORKSPACE_SINCE

The minimal object version supporting this request

REQ_DESTROY_OPCODE

The wire opcode for this request

REQ_DESTROY_SINCE

The minimal object version supporting this request